backend/requirements.txt
# Web Framework
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Scraping
beautifulsoup4==4.12.2
selenium==4.16.0
requests==2.31.0
python-jobspy==1.1.37
lxml==5.3.0

# Data Processing
pandas==2.1.4
python-rapidjson==1.14
rapidfuzz==3.6.1

# AI & NLP
google-generativeai==0.3.2
anthropic==0.7.8
openai==1.6.1

# File Processing
PyPDF2==3.0.1
python-docx==1.1.0
reportlab==4.0.7

# Scheduling
apscheduler==3.10.4

# Notifications
aiosmtplib==3.0.1
python-telegram-bot==20.7

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0

# Code Quality
black==23.12.1
ruff==0.1.9

# Monitoring
psutil==5.9.6
